% From SamplerCompare, (c) 2010 Madeleine Thompson
% $Id: dist-class.Rd 1484 2010-08-24 15:39:22Z mthompson $

\name{dist-class}
\alias{dist-class}
\title{A class representing a probability distribution}

\description{This class represents a probability distribution.
See \code{\link{make.dist}} for more information.}
