% From SamplerCompare, (c) 2010-2011 Madeleine Thompson

\name{twonorm}

\alias{twonorm}

\title{Euclidean norm of a vector}

\description{Computes the Euclidean norm of a vector.}

\usage{twonorm(x)}

\arguments{
  \item{x}{A vector.}
}

\details{
  \code{twonorm} computes the Euclidean norm of a vector:
  \code{sqrt(sum(x^2))}.
}

\value{
  A numeric vector of length one containing the two-norm of \var{x}.
}
